Resolved circular depenencies in groups/interactions
This involved splitting out a couple of functions into their own
files/modules in order to remove the need for any circular dependencies
between files
