+++ src/lib/server/api/people/groups/cache.ts
@@ -1,0 +1,4 @@
+export function redisString(instanceId: number, groupId: number | 'all', banned?: boolean) {
+	const bannedSuffix = banned ? ':banned' : '';
+	return `i:${instanceId}:groups:${groupId}${bannedSuffix}`;
+}
